Removed remaining old-style c function declarations, added more size_ts and compiled almost cleanly with wextra and some other highfalutin compiler warnings.

+++ src/Cuts/pspace_landau.c
@@ -29,7 +29,7 @@
 void
 correct_pspace_landau( struct site *__restrict A ,
 		       const struct veclist *__restrict list ,
-		       const int *__restrict in ,
+		       const size_t *__restrict in ,
 		       const size_t DIMS )
 {
   double ave_err = 0. ;
